1st/requirements.txt
uvicorn
fastapi
sqlalchemy
psycopg2-binary
setuptools 
wheel

